+++ apps/docs/assets/scripts/main.js
@@ -1,11 +1,11 @@
-window.onload = () => {
+window.addEventListener('load', () => {
   addSideDrawerListeners();
-};
+})
 
 const addSideDrawerListeners = () => {
   const sideDrawer = document.querySelector('vwc-side-drawer#sidedrawer');
-  sideDrawer.addEventListener('close', () => { toggleSideDrawerButtonIcon(false); });
-  sideDrawer.addEventListener('open', () => { toggleSideDrawerButtonIcon(true); });
+  sideDrawer.addEventListener('close', () => toggleSideDrawerButtonIcon(false));
+  sideDrawer.addEventListener('open', () => toggleSideDrawerButtonIcon(true));
 }
 
 const toggleSideDrawerButton = () => {
@@ -22,24 +22,6 @@ const codeBlockButtonClick = (button) => {
   const details = button.closest('vwc-action-group').nextElementSibling;
   details.open = !details.open;
   button.ariaExpanded = details.open;
-};
-
-const codeCopyButtonClick = (button) => {
-  const details = button.closest('vwc-action-group').nextElementSibling;
-  const { textContent } = details;
-  navigator.clipboard.writeText(textContent.trim())
-    .then(() => {
-      /* clipboard successfully set */
-      button.icon = 'check-line';
-    })
-    .catch(() => {
-      /* clipboard write failed */
-      button.icon = 'close-line';
-    });
-
-  setTimeout(() => {
-    button.icon = 'copy-2-line';
-  }, 1000);
 };
 
 const onloadIframe = (iFrame) => {
@@ -59,8 +41,8 @@ const iframeObservers = new WeakMap();
 
 const autoResize = (iFrame) => {
   new ResizeObserver((entries, observer) => {
-	if (entries.length === 0) return;
-	iFrame.style.height = Math.max(150, entries[0].contentRect.height) + "px";
+    if (entries.length === 0) return;
+    iFrame.style.height = Math.max(150, entries[0].contentRect.height) + "px";
     clearTimeout(iframeObservers.get(iFrame));
     iframeObservers.set(iFrame, setTimeout(() => {
       observer.disconnect();
@@ -70,6 +52,15 @@ const autoResize = (iFrame) => {
 };
 
 const setCurrentIframeTheme = (toggle, iFrame) => {
-  const theme = toggle.icon === "dark-mode-solid" ? '<link rel="stylesheet" href="/assets/styles/tokens/theme-dark.css" media="all">' : '<link rel="stylesheet" href="/assets/styles/tokens/theme-light.css" media="all">';
-  iFrame.contentWindow.document.head?.insertAdjacentHTML("beforeend", theme);
+  const iframeHead = iFrame.contentWindow.document.head;
+
+  const displayMode = toggle.icon === "dark-mode-solid" ? 'dark' : 'light';
+  const theme = `<link id="theme-link" rel="stylesheet" href="/assets/styles/tokens/theme-${displayMode}.css" media="all">`;
+
+  const themeLink = iframeHead.querySelector('#theme-link');
+  if (themeLink) {
+    themeLink.outerHTML = theme;
+  } else {
+    iframeHead.insertAdjacentHTML("beforeend", theme);
+  }
 }