+++ libs/components/src/lib/alert/index.ts
@@ -1,0 +1,4 @@
+import { registerAlert } from './definition';
+
+
+registerAlert();