+++ libs/components/src/lib/card/card.ts
@@ -5,6 +5,11 @@ import {attr, observable} from '@microsoft/fast-element';
  * Base class for card
  *
  * @public
+ * @slot graphic - The graphic slot overrides the icon property.
+ * @slot media - The media slot is mainly for images or video content above the card header.
+ * @slot meta - The meta slot is for action content in the card header.
+ * @slot footer - The footer slot is for content in the card footer.
+ * @slot main - Assign nodes to main slot to fully override a card's predefined flow and style with your own.
  */
 export class Card extends FoundationElement {