+++ libs/components/src/lib/combobox/combobox.ts
@@ -9,11 +9,12 @@ import { AffixIcon } from '../../shared/patterns';
  *
  * @public
  */
-type PopupPlacement = 'top' | 'bottom';
+export type PopupPlacement = 'top' | 'bottom';
 
 /**
  * Base class for combobox
  *
+ * @slot - Default slot.
  */
 @formElements
 export class Combobox extends FoundationCombobox {