+++ libs/components/src/lib/data-grid/data-grid.ts
@@ -1,9 +1,22 @@
 import { DataGrid as FoundationElement } from '@microsoft/fast-foundation';
+import {attr} from '@microsoft/fast-element';
 
+export const DataGridSelectionMode = {
+	none: 'none',
+	singleRow: 'single-row',
+	multiRow: 'multi-row',
+} as const;
+
+export type ValueOf<T> = T[keyof T];
+
+export type DataGridSelectionMode = ValueOf<typeof DataGridSelectionMode>;
 /**
  * Base class for data-grid
  *
  * @public
+ * @slot - Default slot.
  */
 export class DataGrid extends FoundationElement {
+	@attr({attribute: 'selection-mode'})
+		selectionMode?: DataGridSelectionMode;
 }