+++ libs/components/src/lib/dialog/dialog.ts
@@ -1,5 +1,5 @@
 import { FoundationElement } from '@microsoft/fast-foundation';
-import {attr} from '@microsoft/fast-element';
+import {attr, observable} from '@microsoft/fast-element';
 
 // Make sure we support Safari 14
 let dialogPolyfill: any;
@@ -17,12 +17,16 @@ let dialogPolyfill: any;
  *
  * @public
  */
-type IconPlacement = 'top' | 'side';
+export type IconPlacement = 'top' | 'side';
 
 /**
  * Base class for dialog
  *
  * @public
+ * @slot graphic - Use the graphic slot in order to replace the icon.
+ * @slot body - Use the body slot in order to add custom HTML to the dialog.
+ * @slot footer - Use the footer slot in order to add action buttons to the bottom of the dialog.
+ * @slot main - Assign nodes to the main slot to fully override a dialog’s predefined flow and style with your own.
  */
 export class Dialog extends FoundationElement {
 	/**
@@ -37,6 +41,7 @@ export class Dialog extends FoundationElement {
 	@attr({attribute: 'icon-placement'}) iconPlacement?: IconPlacement;
 	@attr subtitle?: string;
 	@attr headline?: string;
+	@attr ({attribute: 'full-width-body', mode: 'boolean'}) fullWidthBody = false;
 	@attr({attribute: 'aria-labelledby'}) ariaLabelledBy: string | null = null;
 	@attr({attribute: 'aria-label'}) override ariaLabel: string | null = null;
 	@attr({attribute: 'aria-describedby'}) ariaDescribedBy: string | null = null;
@@ -110,7 +115,7 @@ export class Dialog extends FoundationElement {
 	close() {
 		if (this.#dialog.open) {
 			this.#dialog.close();
-			this.dispatchEvent(new CustomEvent('close', {bubbles: true, composed: true, detail: this.returnValue}));
+			this.$emit('close', this.returnValue, { bubbles: false });
 		}
 
 		this.open = false;
@@ -145,4 +150,17 @@ export class Dialog extends FoundationElement {
 		this.#dialog.removeEventListener('click', this.#handleScrimClick);
 		this.#dialog.removeEventListener('submit', this.#handleInternalFormSubmit);
 	}
+
+
+	/**
+	 *
+	 * Slot observer:
+	 *
+	 * @internal
+	 */
+
+
+	@observable bodySlottedContent?: HTMLElement[];
+	@observable footerSlottedContent?: HTMLElement[];
+
 }