+++ libs/components/src/lib/radio-group/radio-group.ts
@@ -5,6 +5,7 @@ import { RadioGroup as FastRadioGroup } from '@microsoft/fast-foundation';
  * Base class for radio-group
  *
  * @public
+ * @slot - Default slot.
  */
 export class RadioGroup extends FastRadioGroup {
 	/**