+++ libs/components/src/lib/slider/slider.ts
@@ -1,5 +1,6 @@
 import { attr } from '@microsoft/fast-element';
 import { Slider as FastSlider } from '@microsoft/fast-foundation';
+import { limit } from '@microsoft/fast-web-utilities';
 
 /**
  * Base class for slider
@@ -24,22 +25,20 @@ export class Slider extends FastSlider {
 	 */
 	override valueChanged(previous: string, next: string): void {
 		if (this.$fastController.isConnected) {
+			// min/max constraints backported from v3
 			const nextAsNumber = parseFloat(next);
-			const value = Math.min(
-				Math.max(this['convertToConstrainedValue'](nextAsNumber), this.min),
-				this.max)
-				.toString();
+			const value = limit(
+				this.min,
+				this.max,
+				this['convertToConstrainedValue'](nextAsNumber)
+			).toString();
 
 			if (value !== next) {
 				this.value = value;
 				return;
 			}
-
-			super.valueChanged(previous, next);
-
-			this['setThumbPositionForOrientation'](this.direction);
-
-			this.$emit('change');
+			// v2 super will still do setThumbPositionForOrientation and emit change
+			super.valueChanged(previous, value);
 		}
 	}
 }