+++ libs/components/src/lib/text-area/text-area.ts
@@ -1,19 +1,20 @@
 import { applyMixins, TextArea as FoundationElement } from '@microsoft/fast-foundation';
 import {attr} from '@microsoft/fast-element';
 import {FormElement, FormElementCharCount, FormElementHelperText, formElements, FormElementSuccessText} from '../../shared/patterns';
+import {ErrorText, errorText} from '../../shared/patterns';
 
-type TextAreaWrap = 'hard' | 'soft' | 'off';
+export type TextAreaWrap = 'hard' | 'soft' | 'off';
 
 /**
  * Base class for text-area
  *
  * @public
  */
+@errorText
 @formElements
 export class TextArea extends FoundationElement {
 	@attr wrap?: TextAreaWrap;
 }
 
-export interface TextArea extends FormElement, FormElementCharCount, FormElementHelperText, FormElementSuccessText{}
+export interface TextArea extends FormElement, ErrorText, FormElementCharCount, FormElementHelperText, FormElementSuccessText{}
 applyMixins(TextArea, FormElementCharCount, FormElementHelperText, FormElementSuccessText);
-