+++ libs/components/src/lib/text-field/text-field.ts
@@ -3,16 +3,17 @@ import {attr} from '@microsoft/fast-element';
 import type {Appearance, Shape} from '../enums';
 import {AffixIcon, FormElementCharCount, FormElementHelperText, FormElementSuccessText} from '../../shared/patterns';
 import {FormElement, formElements} from '../../shared/patterns';
+import {ErrorText, errorText} from '../../shared/patterns';
 
-
-type TextFieldAppearance = Extract<Appearance, Appearance.Outlined | Appearance.Ghost>;
-type TextFieldShape = Extract<Shape, Shape.Rounded | Shape.Pill>;
+export type TextFieldAppearance = Extract<Appearance, Appearance.Outlined | Appearance.Ghost>;
+export type TextFieldShape = Extract<Shape, Shape.Rounded | Shape.Pill>;
 
 /**
  * Base class for text-field
  *
  * @public
  */
+@errorText
 @formElements
 export class TextField extends FoundationTextfield {
 	@attr appearance?: TextFieldAppearance;
@@ -20,7 +21,5 @@ export class TextField extends FoundationTextfield {
 	@attr autoComplete?: string;
 }
 
-export interface TextField extends AffixIcon, FormElement, FormElementCharCount, FormElementHelperText, FormElementSuccessText{}
+export interface TextField extends AffixIcon, ErrorText, FormElement, FormElementCharCount, FormElementHelperText, FormElementSuccessText{}
 applyMixins(TextField, AffixIcon, FormElementCharCount, FormElementHelperText, FormElementSuccessText);
-
-