+++ libs/components/src/lib/tree-view/definition.ts
@@ -1,0 +1,25 @@
+import type { FoundationElementDefinition } from '@microsoft/fast-foundation';
+import { registerFactory } from '../../shared/design-system';
+import styles from './tree-view.scss';
+import { TreeView } from './tree-view';
+import { TreeViewTemplate as template } from './tree-view.template';
+
+export const treeViewDefinition = TreeView.compose<FoundationElementDefinition>(
+	{
+		baseName: 'tree-view',
+		template: template as any,
+		styles,
+	}
+);
+
+/**
+ * @internal
+ */
+export const treeViewRegistries = [treeViewDefinition()];
+
+/**
+ * Registers the tree-view element with the design system.
+ *
+ * @param prefix - the prefix to use for the component name
+ */
+export const registerTreeView = registerFactory(treeViewRegistries);